pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ch9329"
version = "1.2.5"
dependencies = ["pyserial"]
requires-python = ">=3"
authors = [{ name = "Pradish Bijukchhe", email = "pradish@sandbox.com.np" }]
description = "Python module to control ch9329"
readme = "README.md"
license = { file = "LICENSE" }
keywords = []
classifiers = ["Programming Language :: Python :: 3"]

[project.urls]
Homepage = "https://github.com/sandbox-pokhara/ch9329"
Issues = "https://github.com/sandbox-pokhara/ch9329/issues"

[project.scripts]
lsch9329 = "ch9329.tools.lsch9329:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"ch9329" = "ch9329"

[tool.isort]
line_length = 79
force_single_line = true

[tool.black]
line-length = 79
preview = true

[tool.pyright]
typeCheckingMode = "strict"
